+++ app/layouts/default.vue
@@ -3,16 +3,16 @@ section
   navigation-bar
   main
     nuxt
-    ts-footer
+  ts-footer
 </template>
 
 <script lang="ts">
 import { Component, Vue } from 'vue-property-decorator'
-import NavigationBar from '~/components/navigation-bar.vue'
-import TsFooter from '~/components/ts-footer.vue'
+import PageHeader from '~/components/page-header.vue'
+import PageFooter from '~/components/page-footer.vue'
 
 @Component({
-  components: { NavigationBar, TsFooter }
+  components: { PageHeader, PageFooter }
 })
 export default class extends Vue {}
 </script>
