+++ app/pages/contact.vue
@@ -24,7 +24,7 @@ export default class extends Vue {}
 <style lang="sass" scoped>
 .container
   margin: 0 auto
-  min-height: 100vh
+  min-height: 90vh
   display: flex
   justify-content: center
   align-items: center
