+++ app/test/components/page-header.spec.ts
@@ -1,15 +1,15 @@
 import { mount, RouterLinkStub } from '@vue/test-utils'
-import NavigationBar from '@/components/navigation-bar.vue'
+import PageHeader from '@/components/page-header.vue'
 
 const makeWrapper = () => {
-  return mount(NavigationBar, {
+  return mount(PageHeader, {
     stubs: {
       NuxtLink: RouterLinkStub
     }
   })
 }
 
-describe('NavigationBar', () => {
+describe('PageHeader', () => {
   test('is a Vue instance', () => {
     const wrapper = makeWrapper()
     expect(wrapper.isVueInstance()).toBeTruthy()
